+++ NEWS.md
@@ -8,3 +8,5 @@
 # rsvddpd 1.0.1
 - Minor bug fixes and addition of user interruption during iterations with longer execution time
 - Allowing robustness parameters beyond 1, with an issue of warning instead of stopping the execution.
+- Added `cv.alpha` function for finding the optimal robustness parameter.
+- Added different convergence rule based warnings.